+++ src/brightsidebudget/journal.py
@@ -14,25 +14,41 @@ class Account(MutableMapping):
     """
     A dictionary-like object that represents an account.
     The key 'Name' and 'Parent' are always present.
-    Provides 'name' and 'parent' methods for convenience.
+    Provides 'name' and 'parent' attributes for convenience.
     """
     def __init__(self, name: str, parent: str = None, tags: dict = None):
         self._data = tags.copy() if tags is not None else {}
         self._data["Name"] = name
         self._data["Parent"] = parent
-        self._check_name_parent()
+        self._check_all()
 
-    def name(self) -> str:
+    @property
+    def name(self):
         return self._data["Name"]
 
-    def parent(self) -> str:
+    @name.setter
+    def name(self, value):
+        self._data["Name"] = value
+        self._check_name()
+
+    @property
+    def parent(self):
         return self._data["Parent"]
+
+    @parent.setter
+    def parent(self, value):
+        self._data["Parent"] = value
+        self._check_parent()
 
     def __getitem__(self, key):
         return self._data[key]
 
     def __setitem__(self, key, value):
         self._data[key] = value
+        if key == "Name":
+            self._check_name()
+        if key == "Parent":
+            self._check_parent()
 
     def __delitem__(self, key):
         if key in ["Name", "Parent"]:
@@ -46,10 +62,11 @@ class Account(MutableMapping):
         return len(self._data)
 
     def __str__(self):
-        return f"Account({self.name()})"
+        return f"Account({self.name})"
 
     def update(self, *args, **kwargs):
         self._data.update(*args, **kwargs)
+        self._check_all()
 
     def keys(self):
         return self._data.keys()
@@ -68,13 +85,26 @@ class Account(MutableMapping):
         a._data = self._data.copy()
         return a
 
-    def _check_name_parent(self):
-        if not isinstance(self._data["Name"], str):
-            raise ValueError("Name must be a string")
+    def get_dict(self) -> dict:
+        """
+        Return the underlying dictionary.
+        Any changes to the dictionary will affect the account.
+        """
+        return self._data
+
+    def _check_all(self):
+        self._check_name()
+        self._check_parent()
+
+    def _check_parent(self):
         if not isinstance(self._data["Parent"], (str, type(None))):
             raise ValueError("Parent must be a string or None")
         if self._data["Parent"] == "":
             self._data["Parent"] = None
+
+    def _check_name(self):
+        if not isinstance(self._data["Name"], str):
+            raise ValueError("Name must be a string")
         if self._data["Name"] == "":
             raise ValueError("Name cannot be an empty string")
 
@@ -91,7 +121,7 @@ class Account(MutableMapping):
         a._data = d.copy() if copy else d
         if "Parent" not in a._data:
             a._data["Parent"] = None
-        a._check_name_parent()
+        a._check_all()
         return a
 
 
@@ -99,7 +129,7 @@ class Posting(MutableMapping):
     """
     A dictionary-like object that represents a posting.
     The key 'Txn', 'Date', 'Account' and 'Amount' are always present.
-    Provides 'txn', 'date', 'account' and 'amount' methods for convenience.
+    Provides 'txn', 'date', 'account' and 'amount' attributes for convenience.
     """
     def __init__(self, txn: int, date: date, account: str, amount: Decimal,
                  tags: dict = None):
@@ -108,28 +138,60 @@ class Posting(MutableMapping):
         self._data["Date"] = date
         self._data["Account"] = account
         self._data["Amount"] = amount
-        self._cast_txn_date_amount()
+        self._check_all()
 
+    @property
     def txn(self) -> int:
         return self._data["Txn"]
 
+    @txn.setter
+    def txn(self, value: int):
+        self._data["Txn"] = value
+        self._check_txn()
+
+    @property
     def date(self) -> date:
         return self._data["Date"]
 
+    @date.setter
+    def date(self, value: date):
+        self._data["Date"] = value
+        self._check_date()
+
+    @property
     def account(self) -> str:
         return self._data["Account"]
 
+    @account.setter
+    def account(self, value: str):
+        self._data["Account"] = value
+        self._check_account()
+
+    @property
     def amount(self) -> Decimal:
         return self._data["Amount"]
 
+    @amount.setter
+    def amount(self, value: Decimal):
+        self._data["Amount"] = value
+        self._check_amount()
+
     def __str__(self):
-        return f"Posting({self.txn()}, {self.date()}, {self.account()}, {self.amount()})"
+        return f"Posting({self.txn}, {self.date}, {self.account}, {self.amount})"
 
     def __getitem__(self, key):
         return self._data[key]
 
     def __setitem__(self, key, value):
         self._data[key] = value
+        if key == "Txn":
+            self._check_txn()
+        if key == "Date":
+            self._check_date()
+        if key == "Account":
+            self._check_account()
+        if key == "Amount":
+            self._check_amount()
 
     def __delitem__(self, key):
         if key in ["Txn", "Date", "Account", "Amount"]:
@@ -169,7 +231,7 @@ class Posting(MutableMapping):
         """
         if tags is None:
             tags = []
-        return tuple([self.date(), self.account(), self.amount()] +
+        return tuple([self.date, self.account, self.amount] +
                      [self._data[k] for k in tags])
 
     @classmethod
@@ -184,54 +246,109 @@ class Posting(MutableMapping):
                 raise ValueError(f"Missing '{k}' key in posting dict")
         p = cls.__new__(cls)
         p._data = d.copy() if copy else d
-        p._cast_txn_date_amount()
+        p._check_all()
         return p
 
-    def _cast_txn_date_amount(self):
-        if not isinstance(self._data["Txn"], int):
-            self._data["Txn"] = int(self._data["Txn"])
-        if not isinstance(self._data["Amount"], Decimal):
-            self._data["Amount"] = Decimal(str(self._data["Amount"]))
+    def get_dict(self) -> dict:
+        """
+        Return the underlying dictionary.
+        Any changes to the dictionary will affect the posting.
+        """
+        return self._data
+
+    def _check_all(self):
+        self._check_account()
+        self._check_txn()
+        self._check_amount()
+        self._check_date()
+
+    def _check_account(self):
+        if not isinstance(self._data["Account"], str):
+            raise ValueError("Account must be a string")
+        if self._data["Account"] == "":
+            raise ValueError("Account cannot be an empty string")
+
+    def _check_date(self):
         if not isinstance(self._data["Date"], date):
             self._data["Date"] = date.fromisoformat(self._data["Date"])
+
+    def _check_amount(self):
+        if not isinstance(self._data["Amount"], Decimal):
+            self._data["Amount"] = Decimal(str(self._data["Amount"]))
+
+    def _check_txn(self):
+        if not isinstance(self._data["Txn"], int):
+            self._data["Txn"] = int(self._data["Txn"])
 
 
 class BAssertion(MutableMapping):
     """
     A dictionary-like object that represents a balance assertion.
     The key 'Date', 'Account', 'Balance' and 'Include children"' are always present.
-    Provides 'date', 'account', 'balance' and 'include_children' methods for convenience.
+    Provides 'date', 'account', 'balance' and 'include_children' attributes for convenience.
     """
-    def __init__(self, dt: date, account: str, balance: Decimal, include_children: bool = True,
+    def __init__(self, date: date, account: str, balance: Decimal, include_children: bool = True,
                  tags: dict = None):
         self._data = tags.copy() if tags is not None else {}
-        self._data["Date"] = dt
+        self._data["Date"] = date
         self._data["Account"] = account
         self._data["Balance"] = balance
         self._data["Include children"] = include_children
-        self._cast_date_amount()
+        self._check_all()
 
+    @property
     def date(self) -> date:
         return self._data["Date"]
 
+    @date.setter
+    def date(self, value: date):
+        self._data["Date"] = value
+        self._check_date()
+
+    @property
     def account(self) -> str:
         return self._data["Account"]
 
+    @account.setter
+    def account(self, value: str):
+        self._data["Account"] = value
+        self._check_account()
+
+    @property
     def balance(self) -> Decimal:
         return self._data["Balance"]
 
+    @balance.setter
+    def balance(self, value: Decimal):
+        self._data["Balance"] = value
+        self._check_balance()
+
+    @property
     def include_children(self) -> bool:
         return self._data["Include children"]
 
+    @include_children.setter
+    def include_children(self, value: bool):
+        self._data["Include children"] = value
+        self._check_include_children()
+
     def __str__(self):
-        return f"BAssertion({self.date()}, {self.account()}, {self.balance()}, \
-                 {self.include_children()})"
+        return f"BAssertion({self.date}, {self.account}, {self.balance}, \
+                 {self.include_children})"
 
     def __getitem__(self, key):
         return self._data[key]
 
     def __setitem__(self, key, value):
         self._data[key] = value
+        if key == "Date":
+            self._check_date()
+        if key == "Account":
+            self._check_account()
+        if key == "Balance":
+            self._check_balance()
+        if key == "Include children":
+            self._check_include_children()
 
     def __delitem__(self, key):
         if key in ["Date", "Account", "Balance", "Include children"]:
@@ -246,6 +363,7 @@ class BAssertion(MutableMapping):
 
     def update(self, *args, **kwargs):
         self._data.update(*args, **kwargs)
+        self._check_all()
 
     def keys(self):
         return self._data.keys()
@@ -278,14 +396,40 @@ class BAssertion(MutableMapping):
         ba._data = d.copy() if copy else d
         if "Include children" not in d:
             ba._data["Include children"] = True
-        ba._cast_date_amount()
+        ba._check_all()
         return ba
 
-    def _cast_date_amount(self):
-        if not isinstance(self._data["Balance"], Decimal):
-            self._data["Balance"] = Decimal(str(self._data["Balance"]))
+    def _check_all(self):
+        self._check_account()
+        self._check_balance()
+        self._check_date()
+        self._check_include_children()
+
+    def _check_account(self):
+        if not isinstance(self._data["Account"], str):
+            raise ValueError("Account must be a string")
+        if self._data["Account"] == "":
+            raise ValueError("Account cannot be an empty string")
+
+    def _check_include_children(self):
+        if not isinstance(self._data["Include children"], bool):
+            if self._data["Include children"] in ["True", "False"]:
+                self._data["Include children"] = self._data["Include children"] == "True"
+
+    def _check_date(self):
         if not isinstance(self._data["Date"], date):
             self._data["Date"] = date.fromisoformat(self._data["Date"])
+
+    def _check_balance(self):
+        if not isinstance(self._data["Balance"], Decimal):
+            self._data["Balance"] = Decimal(str(self._data["Balance"]))
+
+    def get_dict(self) -> dict:
+        """
+        Return the underlying dictionary.
+        Any changes to the dictionary will affect the balance assertion.
+        """
+        return self._data
 
 
 class BAssertionFail():
@@ -297,14 +441,14 @@ class BAssertionFail():
         """
         The difference between the expected balance and the actual balance.
         """
-        return self.bassertion.balance() - self.actual_balance
+        return self.bassertion.balance - self.actual_balance
 
     def __str__(self):
         return f"BAssertionFail({self.bassertion}, {self.actual_balance})"
 
     def error_msg(self):
-        s = f"Balance of {self.bassertion.account()} on {self.bassertion.date()} is "
-        s += f"{self.actual_balance} instead of {self.bassertion.balance()}. Diff: {self.diff()}"
+        s = f"Balance of {self.bassertion.account} on {self.bassertion.date} is "
+        s += f"{self.actual_balance} instead of {self.bassertion.balance}. Diff: {self.diff()}"
         return s
 
 
@@ -359,13 +503,12 @@ class Journal():
         self.roots: list[Account] = None
         self.postings_by_txn: dict[int, list[Posting]] = None
         self.postings_by_acc: dict[str, list[Posting]] = None
-        # Balances is a dict of dicts. The outer dict is indexed by account
-        # identifier and the inner dict is indexed by date. The inner dict value
-        # is a tuple (flow, balance) where flow is the sum of all postings on
-        # that date and balance is the sum of all postings up to that date. The
-        # inner dict contains all dates between the first and last dates that
-        # are recorded with the inner dict.
-        self.balances: dict[str, tuple[date, date, dict[date, tuple[Decimal, Decimal]]]] = {}
+        # Dictionary of account daily flow and balance. For each account we
+        # store the minimum date, the maximum index where the balance is valid
+        # and a list of tuples with the daily flow and balance. The balance can
+        # become invalid when a posting is added to the account. It will be
+        # recomputed when needed.
+        self.balances: dict[str, tuple[date, int, list[tuple[Decimal, Decimal]]]] = None
         self.bassertions_by_acc: dict[str, list[BAssertion]] = None
 
         self._init()
@@ -374,133 +517,161 @@ class Journal():
         # Verify accounts
         seen = set()
         for a in self.accounts:
-            if a.name() in seen:
-                raise ValueError(f"Duplicate account {a.name()}")
-            seen.add(a.name())
+            if a.name in seen:
+                raise ValueError(f"Duplicate account {a.name}")
+            seen.add(a.name)
         self.accounts_graph = nx.DiGraph()
         for acc in self.accounts:
-            self.accounts_graph.add_node(acc.name(), account=acc)
+            self.accounts_graph.add_node(acc.name, account=acc)
         for acc in self.accounts:
-            p = acc.parent()
+            p = acc.parent
             if p:
                 if p not in self.accounts_graph:
                     raise ValueError(f"Unknown parent: {p}")
-                self.accounts_graph.add_edge(p, acc.name())
+                self.accounts_graph.add_edge(p, acc.name)
         if not nx.is_directed_acyclic_graph(self.accounts_graph):
             cycle = nx.find_cycle(self.accounts_graph)
             msg = " -> ".join([x for x, _ in cycle])
             msg = f"{msg} -> {cycle[0][0]}"
             raise ValueError(f"Cycle in accounts: {msg}")
         for p in self.postings:
-            if p.account() not in self.accounts_graph:
-                raise ValueError(f"Unknown account: {p.account()}")
+            if p.account not in self.accounts_graph:
+                raise ValueError(f"Unknown account: {p.account}")
         for ba in self.bassertions:
-            if ba.account() not in self.accounts_graph:
-                raise ValueError(f"Unknown account: {ba.account()}")
+            if ba.account not in self.accounts_graph:
+                raise ValueError(f"Unknown account: {ba.account}")
 
         # Add useful information to accounts graph
         self.roots = [self.accounts_graph.nodes[n]["account"]
                       for (n, degree) in self.accounts_graph.in_degree
                       if degree == 0]
         for r in self.roots:
-            self.accounts_graph.nodes[r.name()]["root"] = r
-            for n in nx.descendants(self.accounts_graph, r.name()):
+            self.accounts_graph.nodes[r.name]["root"] = r
+            for n in nx.descendants(self.accounts_graph, r.name):
                 self.accounts_graph.nodes[n]["root"] = r
         for n in self.accounts_graph.nodes:
             self.accounts_graph.nodes[n]["depth"] = len(nx.ancestors(self.accounts_graph, n))
 
         # Compute txns_by_id and txns_by_acc
         self.postings_by_txn = {}
-        self.postings_by_acc = {acc.name(): [] for acc in self.accounts}
+        self.postings_by_acc = {acc.name: [] for acc in self.accounts}
         for p in self.postings:
-            if p.txn() not in self.postings_by_txn:
-                self.postings_by_txn[p.txn()] = []
-            self.postings_by_txn[p.txn()].append(p)
-            self.postings_by_acc[p.account()].append(p)
+            if p.txn not in self.postings_by_txn:
+                self.postings_by_txn[p.txn] = []
+            self.postings_by_txn[p.txn].append(p)
+            self.postings_by_acc[p.account].append(p)
 
         # Verify txns
         for k, v in self.postings_by_txn.items():
-            s = sum([t.amount() for t in v])
+            s = sum([t.amount for t in v])
             if s != Decimal("0"):
                 raise ValueError(f"Txn {k} is not balanced. Sum: {s}")
-            dt_count = len({t.date() for t in v})
+            dt_count = len({t.date for t in v})
             if dt_count != 1:
                 raise ValueError(f"Txn {k} has {dt_count} dates")
 
         # Verify bassertions
         seen = set()
         for ba in self.bassertions:
-            if (ba.date(), ba.account()) in seen:
-                raise ValueError(f"Duplicate bassertion: {ba.date()} {ba.account()}")
-            seen.add((ba.date(), ba.account()))
+            if (ba.date, ba.account) in seen:
+                raise ValueError(f"Duplicate bassertion: {ba.date} {ba.account}")
+            seen.add((ba.date, ba.account))
+
+        # Initialize balances
+        self.balances = {}
+        for acc in self.accounts:
+            self.balances[acc.name] = (None, None, None)
 
         # Compute bassertions_by_acc
-        self.bassertions_by_acc = {acc.name(): [] for acc in self.accounts}
+        self.bassertions_by_acc = {acc.name: [] for acc in self.accounts}
         for ba in self.bassertions:
-            self.bassertions_by_acc[ba.account()].append(ba)
+            self.bassertions_by_acc[ba.account].append(ba)
 
-    def _init_balance(self, account: str) -> tuple[date, date, dict[date, tuple[Decimal, Decimal]]]:
+    def _init_balance(self, account: str) -> tuple[date, int, list[tuple[Decimal, Decimal]]]:
         ps = self.postings_by_acc[account]
         if not ps:
-            return (None, None, {})
-        min_date = min(t.date() for t in ps)
-        max_date = max(t.date() for t in ps)
+            return (None, None, [])
+        min_date = min(t.date for t in ps)
+        min_date = date(min_date.year, 1, 1)
+        max_date = max(t.date for t in ps)
+        max_date = date(max_date.year, 12, 31)
 
-        # To make our lives easier, we add all dates to all accounts. Since
-        # there is only 365 days in a year and no one keeps 10 000 years of
-        # financial records, this is not a big deal. The alternative would be
-        # to use a SortedDict
-        xs = dict((min_date + timedelta(days=x), Decimal("0"))
-                  for x in range((max_date - min_date).days + 1))
+        # To make our lives easier, we add all dates. Since there is only 365
+        # days in a year and no one keeps 10 000 years of financial records,
+        # this is not a big deal.
+
+        xs = [Decimal("0") for _ in range((max_date - min_date).days + 1)]
 
         # Compute flow
         for p in ps:
-            xs[p.date()] += p.amount()
+            idx = (p.date - min_date).days
+            xs[idx] += p.amount
 
         # Compute balance
         total = Decimal("0")
-        for d, v in xs.items():
+        for idx, v in enumerate(xs):
             total += v
-            xs[d] = (v, total)
+            xs[idx] = (v, total)
 
-        self.balances[account] = (min_date, max_date, xs)
+        self.balances[account] = (min_date, len(xs) - 1, xs)
 
-        return (min_date, max_date, xs)
+        return (min_date, len(xs) - 1, xs)
+
+    def _recompute_balance(self, account: str) -> list[tuple[Decimal, Decimal]]:
+        (min_date, max_bal_idx, xs) = self.balances[account]
+
+        total = xs[max_bal_idx][1]
+        start_idx = max_bal_idx + 1
+        for idx, v in enumerate(xs[start_idx:], start=start_idx):
+            total += v
+            xs[idx] = (v, total)
+
+        self.balances[account] = (min_date, len(xs) - 1, xs)
+
+        return xs
 
     def check_bassertions(self) -> list[BAssertionFail]:
         err = []
         for ba in self.bassertions:
-            actual_balance = self.balance(ba.account(), ba.date(), ba.include_children())
-            diff = ba.balance() - actual_balance
+            actual_balance = self.balance(ba.account, ba.date, ba.include_children)
+            diff = ba.balance - actual_balance
             if diff != Decimal("0"):
                 err.append(BAssertionFail(bassertion=ba, actual_balance=actual_balance))
         return err
 
     def balance(self, account: str, date: date, include_children: bool = True) -> Decimal:
-        if account not in self.balances:
-            (min_date, max_date, d) = self._init_balance(account)
-        else:
-            (min_date, max_date, d) = self.balances[account]
+        (min_date, max_bal_idx, xs) = self.balances[account]
+        if xs is None:
+            (min_date, max_bal_idx, xs) = self._init_balance(account)
+
         if min_date is None or date < min_date:
-            return Decimal("0")
-        if date > max_date:
-            date = max_date
-        total = d[date][1]
+            total = Decimal("0")
+        else:
+            idx = (date - min_date).days
+            if idx >= len(xs):
+                idx = len(xs) - 1
+            if idx > max_bal_idx:
+                xs = self._recompute_balance(account)
+            total = xs[idx][1]
+
         if include_children:
             for c in self.accounts_graph.successors(account):
                 total += self.balance(c, date, include_children=True)
         return total
 
     def flow(self, account: str, date: date, include_children: bool = True) -> Decimal:
-        if account not in self.balances:
-            (min_date, max_date, d) = self._init_balance(account)
-        else:
-            (min_date, max_date, d) = self.balances[account]
+        (min_date, _, xs) = self.balances[account]
+        if xs is None:
+            (min_date, _, xs) = self._init_balance(account)
+
         if min_date is None or date < min_date:
-            return Decimal("0")
-        if date > max_date:
-            return Decimal("0")
-        total = d[date][0]
+            total = Decimal("0")
+        else:
+            idx = (date - min_date).days
+            if idx >= len(xs):
+                total = Decimal("0")
+            else:
+                total = xs[idx][0]
         if include_children:
             for c in self.accounts_graph.successors(account):
                 total += self.flow(c, date, include_children=True)
@@ -518,7 +689,7 @@ class Journal():
         """
         parents = []
         while account:
-            account = self.accounts_graph.nodes[account]["account"].parent()
+            account = self.accounts_graph.nodes[account]["account"].parent
             if account:
                 parents.append(self.accounts_graph.nodes[account]["account"])
         return parents
@@ -552,16 +723,16 @@ class Journal():
 
         if extra.depth:
             for a in accounts:
-                a[extra.depth_tag] = (self.accounts_graph.nodes[a.name()]["depth"] +
+                a[extra.depth_tag] = (self.accounts_graph.nodes[a.name]["depth"] +
                                       extra.depth_start)
         if extra.hierarchy:
             max_depth = nx.dag_longest_path_length(self.accounts_graph) + 1
             for a in accounts:
-                parents = [a] + self.parents(a.name())
+                parents = [a] + self.parents(a.name)
                 parents.reverse()
                 for i in range(max_depth):
                     if i < len(parents):
-                        x = parents[i].name()
+                        x = parents[i].name
                     else:
                         x = None
                     a[extra.hierarchy_tag_format.format(i + extra.depth_start)] = x
@@ -582,33 +753,33 @@ class Journal():
         else:
             ps_by_id = defaultdict(list)
             for p in ps:
-                ps_by_id[p.txn()].append(p)
+                ps_by_id[p.txn].append(p)
         if today is None:
             today = date.today()
         if extra is None:
             extra = PostingExtraTags()
         if extra.account_tags:
             accs = self.accounts_extra(extra=extra.account_tags_extra)
-            accs_extra_by_name = {a.name(): a for a in accs}
+            accs_extra_by_name = {a.name: a for a in accs}
         ffm = extra.first_fiscal_month
         for p in ps:
             if extra.account_tags:
-                d = accs_extra_by_name[p.account()]
+                d = accs_extra_by_name[p.account]
                 p.update(d._data)
                 del p["Name"]  # Already in the account field
             if extra.future_date:
-                p[extra.future_date_tag] = p.date() > today
+                p[extra.future_date_tag] = p.date > today
             if extra.last_x_days:
-                dt = p.date()
+                dt = p.date
                 for x in extra.last_x_days:
                     p[extra.last_x_days_tag_format.format(x)] = dt > today - timedelta(days=x)
             if extra.year:
-                p[extra.year_tag] = p.date().year
+                p[extra.year_tag] = p.date.year
             if extra.month:
-                p[extra.month_tag] = p.date().month
+                p[extra.month_tag] = p.date.month
             if extra.txn_accounts:
-                txn_id = p.txn()
-                p[extra.txn_accounts_tag] = sorted({x.account()
+                txn_id = p.txn
+                p[extra.txn_accounts_tag] = sorted({x.account
                                                     for x in ps_by_id[txn_id]})
                 if extra.txn_accounts_as_str:
                     p[extra.txn_accounts_tag] = (extra
@@ -616,12 +787,12 @@ class Journal():
                                                  .join(p[extra.txn_accounts_tag]))
 
             if extra.fiscal_year:
-                if p.date().month >= ffm:
-                    p[extra.fiscal_year_tag] = p.date().year
+                if p.date.month >= ffm:
+                    p[extra.fiscal_year_tag] = p.date.year
                 else:
-                    p[extra.fiscal_year_tag] = p.date().year - 1
+                    p[extra.fiscal_year_tag] = p.date.year - 1
             if extra.fiscal_month:
-                p[extra.fiscal_month_tag] = ((p.date().month - ffm) % 12) + 1
+                p[extra.fiscal_month_tag] = ((p.date.month - ffm) % 12) + 1
         return ps
 
     def fingerprints(self, tags: list[str] = None) -> dict[tuple, int]:
@@ -673,15 +844,15 @@ class Journal():
 
 def find_faulty_postings(j: Journal, fail: BAssertionFail,
                          days_limit: int = 7) -> Union[list[Posting], None]:
-    acc = fail.bassertion.account()
+    acc = fail.bassertion.account
     ps = j.postings_by_acc[acc]
-    if fail.bassertion.include_children():
+    if fail.bassertion.include_children:
         for c in j.descendants(acc):
             ps.extend(j.postings_by_acc[c])
-    dt = fail.bassertion.date()
-    ps = [t for t in ps if t.date() <= dt and t.date() >= dt - timedelta(days=days_limit)]
-    ps.sort(key=lambda t: t.date(), reverse=True)
-    subset = subset_sum([p.amount() for p in ps], -fail.diff())
+    dt = fail.bassertion.date
+    ps = [t for t in ps if t.date <= dt and t.date >= dt - timedelta(days=days_limit)]
+    ps.sort(key=lambda t: t.date, reverse=True)
+    subset = subset_sum([p.amount for p in ps], -fail.diff())
     if not subset:
         return None
     else: